# Conflicts:
#	game/src/anki/cozmo/game/comms/cozmoGame.cpp
#	lib/anki/cozmo-engine
#	lib/anki/products-cozmo-assets
#	simulator/controllers/webotsCtrlKeyboard/webotsCtrlKeyboard.cpp
#	simulator/worlds/cozmoViz.wbt
#	unity/ios/CozmoUnity_iOS.xcodeproj/project.pbxproj

+++ project/iosCopyResources.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 set -e
+
 SRCDIR="$1"
 ASSETSRDIR="$2"
 DSTDIR="$3"
@@ -18,9 +19,11 @@ TOPLEVEL=`$GIT rev-parse --show-toplevel`
 cd $TOPLEVEL
 
 SPHINXDIR="$TOPLEVEL"/generated/resources/pocketsphinx
+ASSETS_SOUND_DIR="$TOPLEVEL"/EXTERNALS/cozmosoundbanks/GeneratedSoundBanks/iOS
 DSTDIR_META=$DSTDIR/cozmo_resources/config/
 DSTDIR_SPHINX=$DSTDIR/cozmo_resources/pocketsphinx/
 DSTDIR_ASSET=$DSTDIR/cozmo_resources/assets/
+DSTDIR_SOUND=$DSTDIR/cozmo_resources/sound
 
 # verify inputs
 if [ -z "$SRCDIR" ] ; then
@@ -48,6 +51,11 @@ if [ -z "DSTDIR" ] ; then
 	exit 1;
 fi
 
+if [ ! -d "$ASSETS_SOUND_DIR" ] ; then
+	echo "wwise sound assets not found $ASSETS_SOUND_DIR"
+	exit 1;
+fi
+
 # copy resources
 mkdir -p "$DSTDIR_META"
 rsync -r -t --exclude=".*" --delete $SRCDIR/ $DSTDIR_META/
@@ -59,6 +67,9 @@ if [ -d "$ASSETSRDIR" ] ; then
   rsync -r -t --exclude=".*" --delete $ASSETSRDIR/ $DSTDIR_ASSET/
 fi
 
+if [ -d "$ASSETS_SOUND_DIR" ]; then
+  rsync -v -r -t --exclude=".*" $ASSETS_SOUND_DIR/ $DSTDIR_SOUND/
+fi
 
 echo asset: $DSTDIR_ASSET
 echo meta: $DSTDIR_META
